Troubleshooting: Extract `Prerendered` component
This handles displaying HTML which was rendered server-side. It avoids
the need to do `dangerouslySetInnerHtml` all over the page, and it
provides the CSS to make our standard rendered wikitext look like the
mockup.
