+++ frontend/templates/troubleshooting/index.tsx
@@ -2,13 +2,8 @@ import { DefaultLayout } from '@layouts/default';
 import { DefaultLayoutProps } from '@layouts/default/server';
 import Head from 'next/head';
 import React from 'react';
-import {
-   Box,
-   chakra,
-   Container,
-   Heading,
-   SystemStyleObject,
-} from '@chakra-ui/react';
+import { Box, Container, Heading } from '@chakra-ui/react';
+import Prerendered from './prerendered';
 import { Problem, TroubleshootingData } from './hooks/useTroubleshootingProps';
 
 const renderStyles = {
@@ -72,48 +67,6 @@ function ConclusionSection({ conclusion }: { conclusion: Problem }) {
       </Box>
    );
 }
-
-const Prerendered = chakra(function Prerendered({
-   html,
-   className,
-}: {
-   html: string;
-   className?: string;
-}) {
-   const renderStyles: SystemStyleObject = {
-      '.headerContainer': {
-         display: 'flex',
-         alignItems: 'baseline',
-         marginBottom: 2,
-      },
-
-      '.selfLink': {
-         display: 'none',
-      },
-
-      h3: {
-         fontSize: 'xl',
-         lineHeight: '1.2',
-      },
-
-      'h3,h4': {
-         fontWeight: 590,
-      },
-
-      'h4,h5': {
-         fontSize: 'md',
-         lineHeight: '1.25',
-      },
-   };
-
-   return (
-      <Box
-         className={className}
-         sx={renderStyles}
-         dangerouslySetInnerHTML={{ __html: html }}
-      />
-   );
-});
 
 Wiki.getLayout = function getLayout(page, pageProps) {
    return <DefaultLayout {...pageProps.layoutProps}>{page}</DefaultLayout>;
